review

+++ backend/learn-domain/src/main/java/com/prosper/learn/domain/service/data/CommentDataService.java
@@ -137,14 +137,14 @@ public class CommentDataService extends AbstractDataService<CommentDO, CommentMa
     /**
      * 根据状态获取评论列表（不缓存）
      */
-    public List<CommentDO> getListByState(int state, int limit) {
+    public List<CommentDO> getListByState(byte state, int limit) {
         return commentMapper.getListByState(state, limit);
     }
 
     /**
      * 根据状态分页获取评论列表（不缓存）
      */
-    public List<CommentDO> getListByStatePaginated(int state, long offsetId, int limit) {
+    public List<CommentDO> getListByStatePaginated(byte state, long offsetId, int limit) {
         return commentMapper.getListByStatePaginated(state, offsetId, limit);
     }
 }
